fetch earnings per machine data and pass to the chart
