+++ packages/web-app/src/modules/balance/utils.tsx
@@ -1,6 +1,8 @@
+import { Guid } from 'guid-typescript'
 import { chunk, groupBy } from 'lodash'
 import moment from 'moment'
-import type { EarningWindow } from './models'
+import type { MachineEarningHistory } from '../../api/machinesApiClient/generated/models'
+import type { EarningPerMachine, EarningWindow } from './models'
 
 export const batchEarningsWindow = (earnings: Map<number, number>, batchSize: number): Map<number, number> => {
   const earningHistory = new Map<number, number>()
@@ -34,4 +36,43 @@ export const getEarningWindowsGroupedByDay = (
       : accumulatedEarningsByTheDay.slice(accumulatedEarningsByTheDay.length - numberOfDays)
 
   return chunkedDaysPeriodEarnings
+}
+
+export type BaseKey = string | number
+
+export const getBaseKeyAsGuid = (key: BaseKey | Guid): Guid => {
+  // TODO: Do not force to string and bypass type check (after Kiota is fixed).
+  // https://github.com/microsoft/kiota-typescript/issues/1113
+  if (key instanceof Guid) {
+    return key.toString() as unknown as Guid
+  }
+  return Guid.parse(key.toString()).toString() as unknown as Guid
+}
+
+export const normalizeEarningsPerMachine = (
+  earningsPerMachine: MachineEarningHistory[] | null,
+): EarningPerMachine | null => {
+  let normalizedEarningsPerMachine: EarningPerMachine | null = null
+  if (earningsPerMachine) {
+    normalizedEarningsPerMachine = earningsPerMachine.reduce((earningsPerMachine, machineEarnings) => {
+      if (machineEarnings.earnings) {
+        const machineId = machineEarnings.machine_id?.toString() as string
+        const earnings = Object.keys(machineEarnings.earnings).map((timestamp: string) => {
+          return {
+            timestamp: moment(timestamp),
+            earnings: (machineEarnings.earnings as Record<string, number>)?.[timestamp] as number,
+          }
+        })
+
+        return {
+          ...earningsPerMachine,
+          [machineId]: earnings,
+        }
+      }
+
+      return earningsPerMachine
+    }, {} as EarningPerMachine)
+  }
+
+  return normalizedEarningsPerMachine
 }
